backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
psycopg[binary]==3.2.5
alembic==1.14.1
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
email-validator==2.2.0
slowapi==0.1.9
sentry-sdk[fastapi]>=2.0.0
resend>=2.0.0
stripe>=8.0.0
pytest==8.3.4
httpx==0.28.1
